Experimental handling of page reload.
This is a proposed implementation of handling page reload while user is
interacting with a 3rd party payment app. This patch adds an event
listener named 'paymentresponse' to the window object. This event should
be fired after page reload, if user has authorized the transaction while
Chrome was not in memory.